SERVICE: Quiz creation DTOs completed

+++ Iconic.Domain/Entitites/Courses/Video.cs
@@ -7,8 +7,8 @@ namespace Iconic.Domain.Entitites.Courses
         public string Title { get; set; }
         public string Description { get; set; }
         public string Thumbnail { get; set; }
-        public long Length { get; set; }
-        public long ViewCount { get; set; }
+        public int Length { get; set; }
+        public int ViewCount { get; set; }
         public string Url { get; set; }
         public int LessonId { get; set; }
         public Lesson Lesson { get; set; }
