+++ Iconic.Domain/Entitites/Quizes/SolvedQuestion.cs
@@ -4,12 +4,12 @@ namespace Iconic.Domain.Entities.Quizes
 {
     public class SolvedQuestion : Auditable
     {
-        public long QuestionId { get; set; }
+        public int QuestionId { get; set; }
         public Question Question { get; set; }
-        public long AnswerId { get; set; }
+        public int AnswerId { get; set; }
         public Answer Answer { get; set; }
         public bool IsCorrect { get; set; }
-        public long QuizResultId { get; set; }
+        public int QuizResultId { get; set; }
         public QuizResult QuizResult { get; set; }
     }
 }
